backend/pyproject.toml
[project]
name = "astra-ai-backend"
version = "0.1.0"
description = "FastAPI backend for Astra AI autonomous research agent"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.43",
  "psycopg[binary]>=3.2.10",
  "pydantic-settings>=2.10.1",
  "email-validator>=2.3.0",
  "python-jose[cryptography]>=3.5.0",
  "passlib[bcrypt]>=1.7.4",
  "requests>=2.32.4",
  "beautifulsoup4>=4.13.5",
  "langchain-core>=0.3.76",
  "langchain-community>=0.3.30",
  "faiss-cpu>=1.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.2",
  "httpx>=0.28.1",
  "ruff>=0.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]
